check solutions before moving on

+++ reset.go
@@ -7,9 +7,13 @@ import (
 )
 
 func (cfg *apiConfig) reset(w http.ResponseWriter, r *http.Request) {
+	if cfg.platform != "dev" {
+		returnErrorMsg(w, http.StatusForbidden, "reset forrbidden")
+		return
+	}
 	if err := cfg.db.DeleteAllUsers(r.Context()); err != nil {
 		log.Printf("error deleting users: %v", err)
-		returnErrorMsg(w, http.StatusInternalServerError)
+		returnErrorMsg(w, http.StatusInternalServerError, "internal server error")
 		return
 	}
 	cfg.fileserverHits.Store(0)
